% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actual_vs_pred.R
\name{plot_actual_vs_pred}
\alias{plot_actual_vs_pred}
\title{Scatter plot comparing actuals and predictions}
\usage{
plot_actual_vs_pred(data, xvar, yvar)
}
\description{
Scatter plot comparing actuals and predictions
}
