% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_importance.R
\name{plot_var_importance}
\alias{plot_var_importance}
\title{Bar plot for variable importance in tree based model}
\usage{
plot_var_importance(fit)
}
\description{
Bar plot for variable importance in tree based model
}
