% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_negative.R
\name{smooth_negative}
\alias{smooth_negative}
\title{Smooth out negative values}
\usage{
smooth_negative(vec = NA)
}
\arguments{
\item{vec}{A numeric vector with negative elements}
}
\value{
vector
}
\description{
Spread negative values through other values with sum remaining the same
}
\examples{
\dontrun{
smooth_negative(vec = c(1:24, -50, rep(10, 5), -100))
smooth_negative(vec = c(0,72.36,0,0,0))
}
}
